+++ framework/Furion/DatabaseAccessor/Repositories/EFCoreRepository.cs
@@ -620,6 +620,15 @@ namespace Furion.DatabaseAccessor
         }
 
         /// <summary>
+        /// 判断是否是 MySql 数据库 官方包（更新不及时，只支持 8.0.23+ 版本， 所以单独弄一个分类）
+        /// </summary>
+        /// <returns>bool</returns>
+        public virtual bool IsMySqlOfficial()
+        {
+            return ProviderName.Equals(DbProvider.MySqlOfficial, StringComparison.Ordinal);
+        }
+
+        /// <summary>
         /// 判断是否是 PostgreSQL 数据库
         /// </summary>
         /// <returns>bool</returns>